requirements.txt
fastapi
uvicorn
pydantic
motor
httpx
pytest
flake8
pytest-asyncio
python-dotenv
